refactor: [riscv] Introduce ALU_OPCODE. Move fetch operand to decode stage from execute stage.

+++ arch/riscv/src/isa/abi_name.rs
@@ -2,47 +2,47 @@
 
 #![allow(dead_code, non_upper_case_globals)]
 /// hardwired zero
-pub const zero: usize = 0;
+pub const zero: u32 = 0;
 /// return address
-pub const ra: usize = 1;
+pub const ra: u32 = 1;
 /// stack pointer
-pub const sp: usize = 2;
+pub const sp: u32 = 2;
 /// global pointer
-pub const gp: usize = 3;
+pub const gp: u32 = 3;
 /// thread pointer
-pub const tp: usize = 4;
+pub const tp: u32 = 4;
 /// temporary registers
-pub const t0: usize = 5;
-pub const t1: usize = 6;
-pub const t2: usize = 7;
+pub const t0: u32 = 5;
+pub const t1: u32 = 6;
+pub const t2: u32 = 7;
 /// saved register / frame pointer
-pub const s0: usize = 8;
-pub const fp: usize = 8;
+pub const s0: u32 = 8;
+pub const fp: u32 = 8;
 /// saved register
-pub const s1: usize = 9;
+pub const s1: u32 = 9;
 /// function arguments / return values
-pub const a0: usize = 10;
-pub const a1: usize = 11;
+pub const a0: u32 = 10;
+pub const a1: u32 = 11;
 /// function arguments
-pub const a2: usize = 12;
-pub const a3: usize = 13;
-pub const a4: usize = 14;
-pub const a5: usize = 15;
-pub const a6: usize = 16;
-pub const a7: usize = 17;
+pub const a2: u32 = 12;
+pub const a3: u32 = 13;
+pub const a4: u32 = 14;
+pub const a5: u32 = 15;
+pub const a6: u32 = 16;
+pub const a7: u32 = 17;
 /// saved registers
-pub const s2: usize = 18;
-pub const s3: usize = 19;
-pub const s4: usize = 20;
-pub const s5: usize = 21;
-pub const s6: usize = 22;
-pub const s7: usize = 23;
-pub const s8: usize = 24;
-pub const s9: usize = 25;
-pub const s10: usize = 26;
-pub const s11: usize = 27;
+pub const s2: u32 = 18;
+pub const s3: u32 = 19;
+pub const s4: u32 = 20;
+pub const s5: u32 = 21;
+pub const s6: u32 = 22;
+pub const s7: u32 = 23;
+pub const s8: u32 = 24;
+pub const s9: u32 = 25;
+pub const s10: u32 = 26;
+pub const s11: u32 = 27;
 ///temporary registers
-pub const t3: usize = 28;
-pub const t4: usize = 29;
-pub const t5: usize = 30;
-pub const t6: usize = 31;
+pub const t3: u32 = 28;
+pub const t4: u32 = 29;
+pub const t5: u32 = 30;
+pub const t6: u32 = 31;
